dnf5: Move DownloadCB from context.cpp into separate source
