+++ dnf5/context.cpp
@@ -19,6 +19,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 
 #include "dnf5/context.hpp"
 
+#include "download_cb.hpp"
 #include "plugins.hpp"
 #include "utils.hpp"
 #include "utils/bgettext/bgettext-mark-domain.h"
@@ -258,87 +259,6 @@ void Context::load_repos(bool load_system) {
     print_info("Repositories loaded.");
 }
 
-
-namespace {
-
-class DownloadCB : public libdnf::repo::DownloadCallbacks {
-public:
-    DownloadCB(libdnf::cli::progressbar::MultiProgressBar & mp_bar, const std::string & what)
-        : multi_progress_bar(&mp_bar),
-          what(what) {
-        auto pb = std::make_unique<libdnf::cli::progressbar::DownloadProgressBar>(-1, what);
-        progress_bar = pb.get();
-        multi_progress_bar->add_bar(std::move(pb));
-    }
-
-    int end(TransferStatus status, const char * msg) override {
-        switch (status) {
-            case TransferStatus::SUCCESSFUL:
-                progress_bar->set_state(libdnf::cli::progressbar::ProgressBarState::SUCCESS);
-                break;
-            case TransferStatus::ALREADYEXISTS:
-                // skipping the download -> downloading 0 bytes
-                progress_bar->set_ticks(0);
-                progress_bar->set_total_ticks(0);
-                progress_bar->add_message(libdnf::cli::progressbar::MessageType::SUCCESS, msg);
-                progress_bar->start();
-                progress_bar->set_state(libdnf::cli::progressbar::ProgressBarState::SUCCESS);
-                break;
-            case TransferStatus::ERROR:
-                progress_bar->add_message(libdnf::cli::progressbar::MessageType::ERROR, msg);
-                progress_bar->set_state(libdnf::cli::progressbar::ProgressBarState::ERROR);
-                break;
-        }
-        multi_progress_bar->print();
-        return 0;
-    }
-
-    int progress(double total_to_download, double downloaded) override {
-        auto total = static_cast<int64_t>(total_to_download);
-        if (total > 0) {
-            progress_bar->set_total_ticks(total);
-        }
-        if (progress_bar->get_state() == libdnf::cli::progressbar::ProgressBarState::READY) {
-            progress_bar->start();
-        }
-        progress_bar->set_ticks(static_cast<int64_t>(downloaded));
-        if (is_time_to_print()) {
-            multi_progress_bar->print();
-        }
-        return 0;
-    }
-
-    int mirror_failure(const char * msg, const char * url) override {
-        //std::cout << "Mirror failure: " << msg << " " << url << std::endl;
-        std::string message = std::string(msg) + " - " + url;
-        progress_bar->add_message(libdnf::cli::progressbar::MessageType::ERROR, message);
-        multi_progress_bar->print();
-        return 0;
-    }
-
-private:
-    static bool is_time_to_print() {
-        auto now = std::chrono::steady_clock::now();
-        auto delta = now - prev_print_time;
-        auto ms = std::chrono::duration_cast<std::chrono::milliseconds>(delta).count();
-        if (ms > 100) {
-            // 100ms equals to 10 FPS and that seems to be smooth enough
-            prev_print_time = now;
-            return true;
-        }
-        return false;
-    }
-
-    static std::chrono::time_point<std::chrono::steady_clock> prev_print_time;
-
-    libdnf::cli::progressbar::MultiProgressBar * multi_progress_bar;
-    libdnf::cli::progressbar::DownloadProgressBar * progress_bar;
-    std::string what;
-};
-
-std::chrono::time_point<std::chrono::steady_clock> DownloadCB::prev_print_time = std::chrono::steady_clock::now();
-
-}  // namespace
 
 void Context::download_urls(
     std::vector<std::pair<std::string, std::filesystem::path>> url_to_dest_path, bool fail_fast, bool resume) {
